requirements.txt
python_emacs==0.2.0
